Renamed class Renamed method name Edit import for renamed class Fixed NPE
